Move storage classes to their own package (#203)
