+++ src/main/scala/com.snowplowanalytics.snowplow.micro/storage/InMemoryStorage.scala
@@ -8,9 +8,10 @@
  * OF THE SOFTWARE, YOU AGREE TO THE TERMS OF SUCH LICENSE AGREEMENT.
  */
 
-package com.snowplowanalytics.snowplow.micro
+package com.snowplowanalytics.snowplow.micro.storage
 
 import cats.effect.IO
+import com.snowplowanalytics.snowplow.micro._
 import com.snowplowanalytics.snowplow.micro.model.ColumnStatsResponse
 import io.circe.Json
